+++ test/compile_test.sh
@@ -14,6 +14,13 @@ assertCapturedSuccess() {
   fi
 }
 
+setupJavaEnv() {
+  # like jvmcommon but without 'ulimit -u', which doesn't work on Travis
+  export JAVA_HOME="$BUILD_DIR/.jdk"
+  export LD_LIBRARY_PATH="$JAVA_HOME/jre/lib/amd64/server:$LD_LIBRARY_PATH"
+  export PATH="$BUILD_DIR/.heroku/bin:$JAVA_HOME/bin:$PATH"
+}
+
 createPom()
 {
   cat > ${BUILD_DIR}/pom.xml <<EOF
@@ -133,6 +140,7 @@ testCompileWithoutSystemProperties() {
   assertCaptured "Installing JDK 1.8"
   assertTrue "Java should be present in runtime." "[ -d ${BUILD_DIR}/.jdk ]"
   assertTrue "Java version file should be present." "[ -f ${BUILD_DIR}/.jdk/version ]"
+  assertTrue "system.properties was not cached" "[ -f $CACHE_DIR/system.properties ]"
 }
 
 testCompile()
@@ -144,6 +152,8 @@ testCompile()
   assertCapturedSuccess
 
   _assertMavenLatest
+  assertTrue "system.properties was not cached" "[ -f $CACHE_DIR/system.properties ]"
+  assertContains "system.properties contains the wrong version" "java.runtime.version=1.8" "$(cat $CACHE_DIR/system.properties)"
 }
 
 testCompilationFailure()
@@ -160,8 +170,8 @@ testNewAppsRemoveM2Cache()
 {
   createPom
   rm -r ${CACHE_DIR} # simulate a brand new app without a cache dir
-
   assertFalse "Precondition: New apps should not have a CACHE_DIR prior to running" "[ -d ${CACHE_DIR} ]"
+  mkdir ${CACHE_DIR}
 
   compile
 
@@ -213,6 +223,19 @@ testCustomSettingsXmlWithUrl()
   unset MAVEN_SETTINGS_URL
 }
 
+testCustomSettingsXmlWithInvalidUrl()
+{
+  createPom
+
+  export MAVEN_SETTINGS_URL="https://example.com/ha7s8duysadfuhasjd/settings.xml"
+
+  compile
+
+  assertCapturedError
+
+  unset MAVEN_SETTINGS_URL
+}
+
 testIgnoreSettingsOptConfig()
 {
   createPom "$(withDependency)"
@@ -254,6 +277,8 @@ EOF
 
 testMavenUpgrade()
 {
+    setupJavaEnv
+
     cat > ${BUILD_DIR}/system.properties <<EOF
 maven.version=3.0.5
 EOF