+++ test/test_test.sh
@@ -5,8 +5,14 @@
 
 capture_test()
 {
-  . $BUILD_DIR/.profile.d/maven.sh
-  capture ${BUILDPACK_HOME}/bin/test ${BUILD_DIR} ${CACHE_DIR} ${ENV_DIR}
+  HOME=$BUILD_DIR . $BUILD_DIR/.profile.d/maven.sh
+
+  # like jvmcommon but without 'ulimit -u', which doesn't work on Travis
+  export JAVA_HOME="$BUILD_DIR/.jdk"
+  export LD_LIBRARY_PATH="$JAVA_HOME/jre/lib/amd64/server:$LD_LIBRARY_PATH"
+  export PATH="$BUILD_DIR/.heroku/bin:$JAVA_HOME/bin:$PATH"
+
+  capture ${BUILDPACK_HOME}/bin/test ${BUILD_DIR} ${ENV_DIR}
 }
 
 capture_test_compile()
@@ -31,6 +37,30 @@ createTestPom()
   <dependencies>
 $1
   </dependencies>
+  <build>
+    <plugins>
+      <plugin>
+        <groupId>org.codehaus.mojo</groupId>
+        <artifactId>exec-maven-plugin</artifactId>
+        <version>1.6.0</version>
+        <executions>
+          <execution>
+            <id>echo-string</id>
+            <phase>verify</phase>
+            <goals>
+              <goal>exec</goal>
+            </goals>
+          </execution>
+        </executions>
+        <configuration>
+          <executable>echo</executable>
+          <arguments>
+            <argument>exec-verify-goal</argument>
+          </arguments>
+        </configuration>
+      </plugin>
+    </plugins>
+  </build>
 </project>
 EOF
 }
@@ -41,12 +71,35 @@ test_test_compile() {
   createTestPom
 
   capture_test_compile
-  assertCaptured "Build was not successful" "BUILD SUCCESS"
   assertEquals "Expected captured exit code to be 0; was <${RETURN}>" "0" "${RETURN}"
+  assertCaptured "Build was not successful" "BUILD SUCCESS"
   assertTrue "mvn should be executable" "[ -x ${BUILD_DIR}/.maven/bin/mvn ]"
   assertTrue "mvn profile should exist" "[ -f ${BUILD_DIR}/.profile.d/maven.sh ]"
 
   capture_test
   assertEquals "Expected captured exit code to be 0; was <${RETURN}>" "0" "${RETURN}"
   assertCaptured "Build was not successful" "BUILD SUCCESS"
+
+  unset M2_HOME
+  unset MAVEN_OPTS
+}
+
+test_verify_compile() {
+  createTestPom
+
+  capture_test_compile
+  assertEquals "Expected captured exit code to be 0; was <${RETURN}>" "0" "${RETURN}"
+  assertCaptured "Build was not successful" "BUILD SUCCESS"
+  assertTrue "mvn should be executable" "[ -x ${BUILD_DIR}/.maven/bin/mvn ]"
+  assertTrue "mvn profile should exist" "[ -f ${BUILD_DIR}/.profile.d/maven.sh ]"
+
+  export MAVEN_HEROKU_CI_GOAL="verify"
+  capture_test
+  assertEquals "Expected captured exit code to be 0; was <${RETURN}>" "0" "${RETURN}"
+  assertCaptured "'mvn verify' did not run" "exec-verify-goal"
+  assertCaptured "Build was not successful" "BUILD SUCCESS"
+
+  unset MAVEN_HEROKU_CI_GOAL
+  unset M2_HOME
+  unset MAVEN_OPTS
 }